Update h5writeDataFrame.c
My understanding is that the current approach attempts to avoid strings of length zero. However, by doing so, it prevents R from creating single character columns in a compound dataset (as a single character column automatically becomes a string of length 2). This is causing a problem for the interoperability with languages which have single character data types and attempt to map an H5 compound data type to a struct containing such a data type.